// src/app/types/my/social-links/platform.ts
export enum SocialPlatform {
  ALI_PAY = 'alipay',
  ALI_WANG_WANG = 'ali-wangwang',
  AMAZON = 'amazon',
  APPLE_APP_STORE = 'apple-app-store',
  APPLE_MUSIC = 'apple-music',
  APPLE_PODCASTS = 'apple-podcasts',
  ART_STATION = 'art-station',
  BAI_HE = 'baihe',
  BANDCAMP = 'bandcamp',
  BAO_BAO_SHU = 'baobaoshu',
  BEHANCE = 'behance',
  BILI_BILI = 'bilibili',
  BLUE_SKY = 'bluesky',
  BOSS = 'boss',
  CAMEO = 'cameo',
  CLUB_HOUSE = 'clubhouse',
  DING_TALK = 'dingtalk',
  DISCORD = 'discord',
  DOU_BAN = 'douban',
  DOU_YIN = 'douyin',
  DRIBBLE = 'dribble',
  DROPBOX = 'dropbox',
  DUO_LIN_GUO = 'duolinguo',
  EMAIL = 'email',
  ESTY = 'esty',
  FACEBOOK = 'facebook',
  FEI_SHU = 'feishu',
  GITHUB = 'github',
  GITLAB = 'gitlab',
  GOOD_READS = 'goodreads',
  GOOGLE_PLAY_STORE = 'google-play-store',
  GOOGLE_PODCAST = 'google-podcast',
  HU_PU = 'hupu',
  INSTAGRAM = 'instagram',
  JD = 'jd',
  SHI_JI_JIA_YUAN = 'jiayuan',
  JUE_JIN = 'juejin',
  KICK = 'kick',
  KUAI_SHOU = 'kuaishou',
  LA_GOU = 'lagou',
  LAST_FM = 'last-fm',
  LINKED_IN = 'linkedin',
  LOFTER = 'lofter',
  MA_FENG_WO = 'mafengwo',
  MAI_MAI = 'maimai',
  MASTONDON = 'mastodon',
  MEDIUM = 'medium',
  MEET_UP = 'meetup',
  MO_MO = 'momo',
  ONLY_FANS = 'onlyfans',
  PATREON = 'patreon',
  PDD = 'pdd',
  PHONE = 'phone',
  PINTREST = 'pintrest',
  POSH_MARK = 'poshmark',
  QIN_BAO_BAO = 'qinbaobao',
  QQ = 'qq',
  REDDIT = 'reddit',
  SIGNAL = 'signal',
  SKYPE = 'skype',
  SLACK = 'slack',
  SNAPCHAT = 'snapchat',
  SOUL = 'soul',
  SOUND_CLOUD = 'soundcloud',
  SPOTIFY = 'spotify',
  SUBSTACK = 'substack',
  TANG_DOU = 'tangdou',
  TAN_TAN = 'tantan',
  TAO_BAO = 'taobao',
  TELEGRAM = 'telegram',
  THREADS = 'threads',
  TIKTOK = 'tiktok',
  TUMBLR = 'tumblr',
  TWITCH = 'twitch',
  UNSPLASH = 'unsplash',
  VENMO = 'venmo',
  VIMEO = 'vimeo',
  WE_COM = 'we-com',
  WEBSITE = 'website',
  WECHAT = 'wechat',
  WEI_BO = 'weibo',
  WHATSAPP = 'whatsapp',
  X = 'x',
  XIAN_YU = 'xianyu',
  XIAO_HONG_SHU = 'xiaohongshu',
  XUE_QIU = 'xueqiu',
  YOUTUBE = 'youtube',
  ZHI_HU = 'zhihu',
}